Разбивает модалку, создание большой фото, рендер комментариев по модулям
